+++ src/index.js
@@ -1,10 +1,9 @@
 // @flow
-import React from "react";
-import { Text, View, Platform, Button } from "react-native";
 import { withState, withHandlers } from "recompose";
 import Style from "@jongold/further";
 
 import Junction from "./junction";
+import Counter from "./counter";
 
 type IWithState = {
   counter: number,
@@ -17,12 +16,10 @@ export const state = withState(
   ({ defaultValue }) => defaultValue || 0, // default
 );
 
-type IWithHandlers =
-  & {
-    increment: Function,
-    decrement: Function,
-  }
-  & IWithState;
+export type IWithHandlers = {
+  increment: Function,
+  decrement: Function,
+} & IWithState;
 
 export const actions = withHandlers({
   increment: ({ setCounter }) => () => setCounter(n => n + 1),
@@ -34,19 +31,5 @@ export const counterStyle = Style.of({
   justifyContent: "center",
   alignItems: "center",
 });
-
-export const Counter = (
-  {
-    increment,
-    counter,
-    decrement,
-  }: IWithHandlers,
-) => (
-  <View style={counterStyle.resolve()}>
-    <Text>Counter is at {counter} on {Platform.OS}</Text>
-    <Button onPress={increment} title="Increment" />
-    <Button onPress={decrement} title="Decrement" />
-  </View>
-);
 
 export default Junction().with(state).with(actions).render(Counter);