+++ src/runtime/server/ssr/withMeta.js
@@ -1,0 +1,13 @@
+// @flow
+import Helmet from "react-helmet";
+import type { Head } from "react-helmet";
+
+type IMetaDataResult = {
+  [key: any]: any,
+  metadata: Head,
+};
+
+export default (x: any): IMetaDataResult => ({
+  ...x,
+  metadata: Helmet.renderStatic(),
+});