+++ changelog.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
   - initial repo management and project setup (@delianides)
   - windows universal app platform as build target (@jbaxleyiii)
+  - dotnet core based server setup with SSR and local dev for react (@jbaxleyiii)
 
 ### Fixed